requirements.txt
streamlit
psycopg2-binary
openai
pgvector
